TECH-342: WIP ResizableBox TS migration
